requirements.txt
Flask
requests
azure-cognitiveservices-speech